refactor: move the different foreign in relevant modules
